cbuild: add a profile context manager to template
This allows you to temporarily switch the current build profile
so that any builds then happen with it. This can be utilized when
building crosstoolchains and that kinda stuff, so that they do not
have to manually manage things like CFLAGS and so on; the system
will automatically dig it up from the profile and use it correctly.