cbuild: initial support for cross compiling
This can generate packages for another architecture, but it can't
yet manage build dependencies, so only basic things (e.g. musl)
can be cross built right now. Other things will fail on not being
able to install makedepends into the sysroot.

In order to deal with makedepends in sysroot, we'll need a way to
mask the files already installed by crosstoolchain (probably via
an internal package that provides= the things we want to leave
alone).

Also, a bunch of linting and maybe a cc wrapper to catch explicit
includes of /usr/include and so on will be needed.