main/*: avoid using post_extract hook
Unlike Void, we default cwd during post_extract to builddir instead
of the extracted wrksrc (for consistency, as that allows us to
ensure that pre/post hooks always have the same wrksrc as the main
do_ function, and the wrksrc is unavailable during do_extract).

So instead, it is encouraged to use pre_/post_patch.