cbuild: rework wrksrc handling logic
This renders create_wrksrc useless, so remove it. It also makes
specifying wrksrc useless too in most cases. Additionally it makes
cwd handling more consistent, since now we can use wrksrc as the
cwd during all steps except when build_wrksrc is specified.

Distfiles are now all extracted in a temporary directory, which
is automatically cleaned. Then the following logic is taken:

- If distfile extraction yields a regular file somewhere or several
  directories, the temporary directory is assumed to be the wrksrc
  and is renamed. This can happen either if there are multiple
  distfiles, or if the single distfile archive does not contain
  a toplevel directory.
- Otherwise (extraction yields a single directory), the extracted
  directory is assumed to be the wrksrc and renamed.

In general, this should cover the intended behavior for vast
majority of projects so that the packager almost never needs to
care. In the other minority of cases, it at least ensures that
cleanup can be performed reliably.

This commit also has the side effect that cbuild will now abort
if the wrksrc is dirty at the point of extraction, which can help
prevent some unintended behaviors.