cbuild: drop the sketchy timestamp logic for package regeneration
Also drop the -E switch. This is now default behavior, i.e. if
there already is a package in the repository that has a version
identical to the template version, nothing will get built.

The -f switch can be used to override that - i.e. instead of
forcing package regeneration even when "up to date", it will
force a rebuild even when it already exists in repo.

This should be a much more intuitive behavior for most people,
and it avoids scenarios where the build would run but package
would not get generated in the end.