cbuild: only always skip by default for the "pkg" target
This means that when running e.g. the "build" or "install" target
explicitly, the system will still go through the steps even when
the final package exists. This is a better behavior as it allows
easy testing of the individual build steps without finishing the
packaging and without passing -f (which will repeat steps even
when already finished in builddir) regardless of the repo state.

For the "pkg" target, the skipping behavior is what we want by
default, as some kind of "dry run" for apk generation does not
make much sense.