cbuild: enforce static cbuild user with uid/gid 1337
Instead of initing passwd/group once and using the info from the
host system, set a static user and group id 1337 with the name
cbuild, and use that by default unless forced (which is typically
when something needs to pretend to be root).

This should be more reliable, and also prevent autodeps from
leaving traces of their created users and groups.