cbuild: rework handling of hardening options for convenience
Previously there were defaults and then profiles could enable
or disable additional hardening options along with the template.
This is inconvenient, as it means an option is always enabled
or disabled, which means a template cannot request a hardening
option only when supported by the target.

This introduces separation between default requested hardening
and target-supported hardening, with profiles only advertising
which hardening options they support (or do not support). That
means if a template requests a hardening option but it comes out
unsupported by profile, it will not be used.

This will later allow us to do things like templates specifying
they want CFI, but having them compile with CFI only on the few
targets that support CFI in Clang.