main/clang-rt-crt-cross: fix build
This is fallout after the wrksrc reorg, it was missed that this
template has multiple distfiles. Ensure to apply patchs in the
proper place, as well as update how things are moved around.

While at it, make cbuild always pass an absolute path to patches
when applying them. This allows the -d flag to work in a way users
would usually expect, rather than requiring moving the patches to
some other place first.

Fixes https://github.com/chimera-linux/cports/issues/44