cbuild: add hook to scan for pkgconf .pc files
Packages are now able to provide `.pc` files in their apk metadata.
If a package contains `.pc` files, its template needs to have a
`pkgconf` hostdep in order to be able to scan the version.

The scanning can be disabled with a special option, it is also
always disabled during stage0 bootstrap pass.