cbuild: automatically resolve symlink dependencies
If either the main package or a subpackage of it provides a
symlink that does not resolve within the package, other subpackages
or the main package are scanned and if it would resolve with that
package installed, a dependency is automatically created.

This allows for much nicer handling of devel packages depending
on their main package.