cbuild: clean up relevant builddir/destdir before pkg by default
Previously the build would continue from where it left off. This
is useful for failed builds, but is also bug-prone. Therefore,
clean up by default, and do not do so only if explicitly requested.

While at it, rename the -I flag to -D, and make it both skip the
dependencies as well as not clean up the dirs. It usually does not
make sense to skip installation of dependencies while actually
doing a build, except in this case.

Also, alter stage 0 bootstrap to be always treated as dirty in
order to allow for proper resumption. From stage 1 onwards, the
build behaves normally (if you want to resume it on failure, you
can use just normal pkg on the relevant failed template, set the
dirty argument and point bldroot/packages to their respective
stage).