cbuild: add support for --force-check
This enforces running tests even when the `!check` option is given
in the template. This is mostly useful for manual testing when
the test suite normally fails or when it's needed to manually
resolve a dependency cycle with checkdepends enabled.

It is not a config file option or anything else, because it should
never be used as a default thing, only when explicitly building
a package.

It is not guaranteed to work, since various templates are not
set up for testing, are missing checkdepends or even have invalid
checkdepends (for things not packaged yet), so it should only be
used for specific packages.