cbuild: use a different method for avoiding backups with BSD patch
This would normally be portable to GNU too, except GNU explicitly
checks if the suffix is empty and errors out, unfortunately.

However, it at least deals with issues when the files have read
only permissions, as this will trigger an unlink() on the target
file. This is probably a bug in BSD patch, since it should patch
the file correctly even with `-V none`, but this is a satisfactory
fix, and we can fix the `-V none` semantics in the tool later.