cbuild: add new exec_wrappers meta variable
This is an easy to use interface to expose some tools into the
executable path under different names. Primarily this would be
used to e.g. expose `gsed` as `sed` for things where the build
system or test suite requires so. It does not affect the generated
package in any way.