main/freetype: build with harfbuzz support
This necessitates a new freetype-bootstrap package, which is
a minimal build of freetype that is used for build of harfbuzz
and its dependencies (cairo and fontconfig). Then once we have
a harfbuzz build, we can build the real freetype, this time with
harfbuzz support.

This results in a shlib dependency cycle, as libharfbuzz will now
depend on so:libfreetype.so.6, while freetype will depend on
so:libharfbuzz.so.0. This is okay though, they will just both
get installed.

The provider_priority ensures that even if freetype-bootstrap and
freetype have the same version, the real freetype is picked up
as a virtual provider of the shlib (and pc file) by default,
except when not yet built.