cbuild: sort of steal xbps-triggers system from void
There is now a system similar to xbps-triggers for generated
scriptlets. The difference is in the organization (there is no
separate UPDATE etc.) as well as in the scope of usage (in void,
these handle the overall trigger functionality, while apk has
real triggers, so we can use those where appropriate and leave
this only for things that insist on doing stuff before/after a
package is installed/updated/removed, or when the event is not
invokable by triggering on a path).

For now, handle system-accounts as a proof of concept, and use
it in dbus.