main/linux: prune initramfs for the kernel when installing
This ensures that the initramfs is not present when installing
a new kernel package. The future kernel trigger will then
regenerate initramfs for all kernels that are missing it.

In the future we should add backup logic to save kernels before
updating or removing them (possibly with some more elaborate
logic in order to not pollute /boot too much) but for now that
is unnecessary.