cbuild: add update-check support
For now this is pretty bare, and likely buggy. The pattenrns and
stuff are adapted from Void's update-check code, since it's too
much of a pain to write from scratch.

Python verbose regexes are used for readability, and it's a bit
more customzable than the xbps-src version, and does not add any
extra dependencies, using largely just Python stuff.