main/libffi: rename to libffi8
Since libffi is subject to soname bumps, we need a way to do
a graceful rebuild. As libffi needs to be installed within the
build root (because of llvm), it would block rebuilding things
after libffi upgrade (as the build root libffi with old soname
would block installation of the new soname and therefore of
libffi-devel).

We can deal with this by versioning the package name (e.g. when
libffi8 becomes libffi9, it will not block installation of both
libffi8 and libffi9 in the build root while everything is being
rebuilt).