cbuild: implement rudimentary staging behavior
This does not actually detect shlib bumps yet, which is the actual
goal of this effort, but it does ensure packages are first stored
in a staging repo and this repo is then merged with the primary
in one go. This affects batching: when a tree of packages is built,
the stage is only cleared at the very end, when the final package
of the batch is built.

In addition to that, the system now allows users to stage packages
explicitly. You can do that either by using the --stage flag or
by using the config file option. This will have various uses on
package builder infrastructure later.

Staging does not affect makedepends. The stage repos are considered
for dependencies, so this behavior is unaffected. However, it does
affect explicit chroots as well as any explicit installations of
packages from the built repos, since the stage repos will not be
considered for that.

Since the actual soname checking work will be a bit more involved,
it is left for another commit. This just prepares the groundwork.