cbuild: implement staging checks
This pile of cursed code serves the purpose of doing proper
staging checks. That means when a rebuilt package changes
virtual providers (usually this means a soname bump), the system
becomes aware of it, and does not unstage the repos until all
packages formerly depending on the old soname have been rebuilt.

You will get a list of stuff to rebuild printed out.

This implementation is not completely reliable, since it does
not care about things moving across repos (so it will not catch
a case when a virtual provider moves from main to contrib, and
there are still revdeps in main) and it will also silently accept
unstaging when there is a potential alternative provider present
in the repository without doing version constraint checks for
each package again. These are tradeoffs made so that the code
can still be reasonably fast and readable. People are expected
to properly check their stuff by hand when doing invasive repo
changes, and possibly force-stage the repository as needed.

We technically don't even really need staging in the first place
since apk handles these things relatively gracefully. However,
it's convenient to be able to be aware of what needs rebuilding
where, as well as to ensure binary repos work smoothly and do
not break users' ability to update, even temporarily.

There may also still be bugs in this implementation. They will
get ironed out over time.