cbuild: support CPPFLAGS as legitimate tool_flags
Now they are exported in the same way as C(XX)FLAGS and so on.
This allows things that directly call the preprocessor to use
correct flags during building, affecting e.g. building of strace
for cross (as it directly calls the preprocessor and needs the
sysroot passed to pick up correct includes).

Closes https://github.com/chimera-linux/cports/pull/61