main/llvm: remove unneeded patch, disable multiarch triple for musl
The removed patch was only relevant to identification of GCC
installs for runtime, which we do not want as clang is our native
compiler and runtime and even if we get gcc at some point, we can
configure it in a way that matches our desired triples.

But while at it, add a patch that kills the multiarch triple
heuristics for musl target triples, as the detected target triples
for multiarch are always gnu ones, and for some archs (e.g. riscv)
it is not defined and it will confuse stuff like python configure
script. A non-multiarch gcc toolchain will print an empty string,
let's ensure we also do that.

Fixes https://github.com/chimera-linux/cports/issues/63