main/dinit-userservd: new package (0.1.0)
This is a combo of a daemon, PAM module and a service that manages
user instances of dinit automatically. It will launch a new dinit
instance when the user first logs in and shut it down once the
last instance of the user has logged out.

It also allows for multiple systemwide user service directories
so things can install custom services meant to be launched by
user instances, and the user simply needs to enable them.