main/dbus: add systemwide user service as proof of concept
For now this is missing some things (logging, and it cannot open
a readiness socket, since dinit does not yet have a way to
substitute variables in paths in services; this will be needed
to shove the user ID and/or homedir in there for logfile and so
on).

But for now this at least works, and we use pipefd for readiness.