main/dinit: implement a workaround to allow wider var substitution
This allows for env var subst in a few more variables, which will
allow us to do things such as refer to $HOME and $UID and so on
in log/socket paths, which enables more flexible user services.

This is very rudimentary and more or less a workaround until a
real solution can be coordinated with upstream.