main/xwayland: provide xserver-common
Once we have a non-xwayland xorg packaged, this will be provided
as a real package, and if xwayland gets installed, it will
transparently replace it. That way we can ensure that the
protocol file is always installed but also not require xorg
for xwayland build and the other way around.