main/*: replace the stupid _mver hack
This has become widespread enough to be an antipattern, and it
does not actually help anything. While removing a fixed-length
suffix may seem hacky, it's the leanest way to do it and in
practice any error made when updating will immediately show
itself (as the URL will fail to fetch).