cbuild: ensure autopkg ver matches base ver in prune-removed
This catches cases such as when a package gets updated, becomes
arch-independent, and no longer has a debug package. Previously,
the old dbg package would linger in the repo because its base
template still existed. We make use of the fact that autopackages
always have a version matching the base package version when they
get generated, so if we catch an autopackage and don't find a base
package of the same version, it's safe to prune it.