main/dinit-chimera: include service targets
The service suite now comes with standard targets early, init,
network, login, boot.

The boot target is the last to come up in a full multi-user mode.

The login target is invoked by gettys, and starts whatever services
are necessary for console login to come up.

The network target is not invoked automatically, but it starts
any services that are required for full networking to come up
(e.g. a dhcp client) and other things that require network to be
started can depend on it without having to depend on any specific
srevices.

The init target is what most services must depend on and is a
prerequisite of login. It itself depends on the early target,
whose sole role is to bring up all the early oneshots - it in
general does not actually start any long-running services. The
init target can start daemons that are required to be up before
anything else and are not transitive dependencies of other things.