main/dinit-chimera: ensure /tmp/.(X11|ICE)-unix is created properly
This makes sure that these directories are root-owned and have
correct permissions, since our xorg/xwayland run without root or
suid and therefore cannot ensure so. This would previously result
in those dirs being created with the owner being the user who first
started it, which is problematic with e.g. GDM and gnome wayland
sessions (GDM creates it first, mutter would fail to start for the
logged in user afterwards because mutter sanitizes the permissions
of that when starting xwayland).