main/xserver-xorg: new package (1.0)
This is a metapackage that provides a distribution of X that
consists of xserver-xorg-minimal (core + libinput driver + xauth
and xinit) plus a suite of utilities that are of practical use
and suitable for most installations.

It does not include anything depending on X Athena Widgets, as
those are not packaged and will not be packaged in main/.

It does not ship additional input or video drivers, as most people
will not need them (libinput is a generic solution for input
drivers, and modesetting will cover most cases for video). People
needing those extra drivers will have to install them separately.

It also does not require any of the X fonts, as those are not
packaged and likely will not be in main/; they are mostly only
useful for legacy applications and portable stuff will use modern
scalable fonts. They may be packaged separately for contrib/ at
some later point.