cbuild: transition to apk3
This requires a very recent Git snapshot apk-tools now.

Keep in mind that since apk3 is not ready yet, the functionality
here is subject to change as well as updates for potential changes
in apk itself. It is every current user's responsibility to ensure
they have a version of apk that matches what is expected by cbuild.

This is for now a work in progress, and some features may not be
reimplemented yet. Also, new features from apk3 have to be adopted
correctly.