cbuild: add bulk-pkg command for bulk builds
This allows one to build several packages instead of just one,
while ensuring that the repo is staged for the whole duration
of the batch build, and being able to record the status of each
package in the batch on a separate file descriptor (which can be
redirected into a file, etc.).

It supports both "march on with more packages after a failed
build" mode (which is the default) and "abort upon first failure"
(which is an option). Most importantly, it will sort the given
set of packages in a way that ensures as little recursion as
possible (i.e. if you have A and B where B depends on A, A will
be built first).

This feature is not finalized yet so it is not documented at this
point. It will be getting more work.