main/lib{unwind,cxx,cxxabi}-cross: consolidate under one template
Thanks to LLVM_ENABLE_RUNTIMES, we can now build those three
components together and save some time and effort. This still
requires the crt+builtins and libc to be built ahead of time,
so that stays as it is. Likewise, compiler-rt sanitizers will
be built separately for the time being.