main/llvm-bootstrap: new package (14.0.1)
This is a "bootstrap" toolchain that is used for stage 2+ builds
of LLVM. Its purpose is to have a throwaway toolchain that we can
compile the final LLVM with. This ensures that LLVM upgrades have
a chance at reproducibility; we can ensure that the toolchain used
to build the final LLVM is always the same as far as its output
goes (in the other case, each LLVM rebuild would be done with the
previously rebuilt one, which is not great for subsequent builds).

Additionally, this ensures that e.g. when we introduce LLVM 15
into the repo, it's all built with LLVM 15 rather than LLVM 14
at first. While this mostly does not matter (the runtimes like
compiler-rt and libcxx are all built with the in-tree clang) it
is still a good thing to have.

Obviously, this toolchain is not used for cross-builds of LLVM
because that is always guaranteed to be built with an identical,
host version of itself (can't cross-compile packages without
having an up to date host toolchain first).

This should be kept at the same version as llvm itself.