cbuild: rework handling of cross toolchains and target autodeps
Instead of installing the cross-toolchain when not installed and
keeping it persistently, simply make it a part of the hostdeps.

Then, before installing hostdeps, just remove the sysroot for the
target entirely and let the hostdeps install cleanly, then set
up the dummy provider on top of that and then install target deps.

There is no need to do the void-style persistent cross-toolchain
anymore because apk is much faster and this adds pretty much no
real overhead, and this is much simpler.