cbuild: defer source_date_epoch calculation to build phase
This dramatically improves parse speed of templates. Since the
setup_reproducible method calls into git and does some pretty
expensive stuff there, we don't want to run it each time a
template is parsed, especially if the calculated values are
only really used during build.

This finally gets rid of the annoying lag after a build is
triggered but before makedepends are installed.