cbuild: rework bulk builder for properly recursive sorting of deps
Previously the code had a sorting problem, because it was not
handling deps fully recursively. Consider you have a package A
which depends on B, which depends on C. You request a bulk build
of A and C. The system attempted a topological sort, accounting
for A's build dependencies as well as C's build dependencies.

However, this shallow sort does not know that B depends on C,
only that A depends on B. Therefore, it could easily happen that
A would attempt building before C, even though C is a "deep"
dependency of A through B.

The sort-dependencies system in xbps-src has a similar problem,
and cbuild inherited it from there. However, since we can afford
to parse templates deeply (parsing every single template at once
in cbuild takes about 1.5 seconds on a computer with a fast CPU
and SSD at the current count, while it would take many times more
than that in xbps-src), we can just do that and fix the whole
issue. It can technically result in a lot of templates being
parsed, but it's not a big deal.