cbuild: only attempt a package in bulk build if all its deps parse
In the hypothetical case of a template of an either direct or
indirect dependency of a package failing to parse at all, this
would result in the dependency graph no longer being complete.

That would mean potential ordering issues related specifically
to that package, as it could build earlier than it should. Stay
on the safe side and skip it in those cases. When that happens,
a fix in the repository with a properly set up builder should
result in a later re-attempt.