cbuild: rework cycle checking around topological sorter
This is a major code simplification, and results in reliable
cycle checks. It does not support cross-compiling, i.e. it parses
templates in non-cross mode.

Previously there was fairly complicated custom in-place setup
that did not work very well. Instead of doing that, just do the
same thing the bulk builder code does, and construct an acyclic
graph for all packages we are checking. Then we can prepare a
sort, and let graphlib find the cycles.

This can only find one (random) cycle at a time. Usually, the
tree should be free of cycles and the cycle check should succeed
with no output.