main/bsdutils-tiny: new package
This is like bsdutils, but installed in a dedicated standalone
prefix and minimized. That means it does not have any of the
dependencies of regular bsdutils (fts, rpmatch and libxo are
statically linked in, color ls is disabled so no ncurses,
libcrypto is disabled so no openssl, libedit is disabled)
other than libc(++).

This makes it suitable for inclusion into early/embedded
environments (in our case initramfs, together with the klibc
utilities that are yet to be added).