cbuild: add a custom fakeroot launcher
This should be a lot less janky, as it does not override the
LD_LIBRARY_PATH variable and generally does very little other
than launching faked, trapping it, and then running the desired
process with libfakeroot in preload.

It's only really used for in-chroot stuff, but that is what
matters the most. For outside-chroot stuff we cannot use it
as we have no control over the external fakeroot installation
and therefore have no way to locate the preload library. But
outside-fakeroot is only really used for bootstrapping bldroots
and such.