cbuild: implement support for separate-directory staging
This makes it possible to stage new packages outside the repo
path instead of inside. That allows one to have e.g. a repository
that is served over http, and a stage repo that is not served at
all, and that way ensure that newly built packages are not visible
to people using the repo until unstaged, even via .stage paths.