cbuild: initial rework of cross sysroot management
This may break recursive cross builds for now. The idea is to set
up the cross sysroot ahead of time, because we need to be able to
refresh remote repos to install target dependencies, and we need
to refresh those for the target architecture, and in order to do
that we need to provide apk with a semi-valid root with the right
architecture set.

We could not do that before because the sysroot was set up after
dependency availability checks were performed.