main/linux-rpi: new package (5.15.x)
This is a 64-bit kernel for Raspberry Pi 3 and 4. Its config is
adapted from Ubuntu, with tweaks to bring it closer to our vanilla
kernel. It is made sure that the structure it follows is identical
to a regular kernel. The other support packages will take care of
setting it up with the firmware.