cbuild: work around cross dummy package hijacking dep resolution
As it is, `apk search` will always consider locally-installed
packages first before repository ones, with no option to override
this. While this is usually harmless, for crossbuilds it results
in the dummy cross package being considered first, which breaks
the availability check when a dependency on some package that
is provided by it is specified.

We will need to find a proper solution for this later as it may
result in other edge cases. That will require at least making
sure that installed packages are ignored, plus some special
handling for virtuals.