cbuild: use exact versions for base-cross-target-meta providers
Instead of using dummy versions with high numbers, provide the
exact versions the "regular" packages use. We can use the versions
from the template as the packages should always be in sync with
the template.

This allows installation of e.g. musl-devel-static into the target
environment when cross-compiling. Previously this was not possible
as the autogenerated -static packages always depend on the -devel
packages with the exact same version as them.

Fixes https://github.com/chimera-linux/cports/issues/120