main/llvm: update to 15.0.2, drop execinfo, enable atomic builtins
We now no longer need libexecinfo in bootstrap path, or have it
be a dependency of llvm or have it be in the container.

This disables gwp-asan so we will figure out how to enable it
again later.

However, it means we no longer need libexecinfo in the cross
toolchain, which simplifies things quite a bit.

Also enable atomic builtins which were accidentally disabled as
llvm disables them by default because gcc setups have libatomic.
We don't have libatomic so we want to have atomic fallbacks for
when e.g. C11 atomics are used with 16-byte values.