cbuild: do not use fakeroot for stage 0 autodep removal
Technically not necessary, as no permissions are being updated
during removal. This could theoretically affect packages that
have really strict permissions (so that root is needed to access
them, i.e. when not even the owner has them) but we control the
packages that are built and installed during stage 0.