main/elogind: add elogind-meta, split pam_elogind
The elogind-meta serves as a "recommends" package for elogind,
which things that expect elogind to be installed but do not
strictly depend on it can depend on. The elogind package will
then get installed through a soft reverse dependency, allowing
people who do not want it to eliminate it (unless something else
has a hard dependency on it).