main/musl: implement a custom exclusive scudo registry
This allows us to have a dedicated TSD per thread, drastically
improving allocator scalability (it will no longer be capped by
a fixed number of threads). The approach taken here is mapping
64 kilobyte-sized chunks (16 pages on most systems, but it can
be as low as 1 e.g. on some ppc64 or aarch64 systems) and giving
out pieces. When we run out (which may happen when we have a lot
of threads), another chunk is mapped. When threads exit, they
donate their memory back to the registry, so that the TSD can
be reused instead of giving out a fresh one. That means we will
only ever have as many chunks as needed to satisfy concurrently
running threads.