cbuild: introduce various hardening options
Minimal UBsan is used by default, with checks for signed integer
overflow, bit shift overflows and integer division by zero.

Intel CET (-fcf-protection=full) is on by default on x86 targets,
while equivalent aarch64 -mbranch-protection=standard is on by
default on aarch64, which appears to be equivalent to just pac-ret
(with hint space instructions for backwards compat) as BTI cannot
be used before armv8.5-a.

Additionally we have Clang CFI and SafeStack. CFI is not on by
default yet (but it will be on supported archs) and SafeStack is
opt-in as it cannot be applied to DSOs (but we can hand-pick some
programs to be compiled with SafeStack, e.g. our core userland
tools).