main/llvm(-bootstrap): enable libcxx assertions, disable cfi/ubsan
These need to be disabled for the toolchain build, because this
template is responsible for building the runtime bits that things
using this hardening link against. We don't currently have any
sort of fine-grained control to build it e.g. for libllvm only,
but this should not be a big loss.