main/*, contrib/*: add temporary tags disabling hidden visibility
This is so that we can enable it by default and CFI along with
it. Metapackages already do not have these added. After some more
testing, CFI and hidden visibility will be turned on out of box,
and after that we will gradually remove these temporary tags after
testing each one.