cbuild: implement repository and sources locking
In this scheme, cbuild locks sources for downloads with a global
lock, which may seem too coarse at first, but when you consider
that multiple templates may share sources (due to same checksum),
there isn't really a better way to do this without complicating
things too much (for a case that is generally unlikely to happen).

As for repos, every target repo path (i.e. what you specify on
cbuild command line) has a per-architecture lock. The lock is
taken for the duration of package generation and staging (i.e.
taken before packages are generated, and released after they are
staged) and then again taken for unstaging. This effectively means
that packages can no longer race each other for getting indexed.

These are advisory locks so they are only respected by cbuild,
but that is fine.