cbuild: use lockf instead of flock
In Python, the lockf interface is pretty much identical to the
flock interface, except POSIX region locks are always used. This
is what we want because we want things to work between mixed local
and NFS mounts (flock on Linux already works with NFS mounts by
using POSIX locks in those cases, but e.g. a server where the fs
is local will not properly interact with a client mounting over
NFS, as the server-local instance will use flock)