cbuild: remove bootstrap.sh
This script has been largely redundant for a while now, as it
uses a self-hosted environment to do the thing. There is little
to no difference between bootstrapping a stage 0 container from
packages and then continuing the bootstrap process from source
and using the script. It also does not work as is anymore anyway
because we have dropped the fakeroot host dependency.