cbuild: implement policy package for autoinstallation of -devel
There is now a new base-devel package that controls this policy.
Having this installed declares that the user wants all development
files for base packages they have to be installed.

This is implemented with a pre-pkg hook; this hook only applies to
-devel packages, by scanning said packages for local dependencies
(i.e. dependencies to other subpackages of the same template) and
generating an implicit install_if list with them.

This hook only does anything if there isn't already a non-empty
install_if list for the package, and if the `scandevelif` option
is not set.

This policy should be fairly conservative, as a devel package may
depend on multiple other subpackages, and will only auto-install
if all of them are already in place. We want this to be a just
a minor convenience though, without overdoing it (any template
that does not work well with this behavior can always special-case
the list).