main/webkitgtk(4): restore functionality
It seems here we are not encountering an UBSan violation in the
webkit code, but rather some kind of strange check artificially
emitted by the compiler, of a different type than we actually
enable in the first place. But this needs more investigation at
a later point. For now just disable the instrumentation.

Also, disable CFI for libwpe as it does not appear to be ready yet.