cbuild: refactor handling of LTO in tool flags
This should also fix stage 0/1 bootstraps for templates that
enable CFI as LTO is not yet available in this stage.

Additionally, it moves disabling LTO for riscv64 to a place where
it's more visible and actually makes sense, as opposed to just
doing it randomly in template.py.

Fixes https://github.com/chimera-linux/cports/issues/141