main/console-setup: make XKB keyboard handling optional
The XKB support in console-setup utilizes a massive Perl script
to do the conversion. This not only takes time on boot (requiring
the caching system) but also pulls in Perl to systems that would
not need it at all otherwise. Since console-setup already supports
regular console keymaps for most part, patch things to do that
by default, loosen the rules a bit, and put the XKB support in its
own package.