main/console-setup: drop guess behavior for consoles and charmap
Instead, always use fixed values in /etc/default/console-setup.

The ACTIVE_CONSOLES guess behavior never worked and it would
always default to tty1-6, as the files it used to guess do not
exist. The CHARMAP guess behavior relied on locale, so it only
worked with glibc.

We now also use ACTIVE_CONSOLES to specify which gettys to start
by default, so it's good to have a reliable value.