main/console-setup: handle non-xkb keymaps for initramfs
When setupcon generates a tiny setup script for initramfs, it will
resolve all absolute paths for loadkeys and copy stuff over. But
when using a keymap name in KMAP, this will not work because it
is not a file path and loadkeys resolves it. We also cannnot use
the parse mode of loadkeys to simply resolve it and pass it,
because it may include other keymap files. Therefore, make a
dump of the keymap instead and include it in initramfs.