main/chimerautils: fix whereis with mandoc environments
The utility was relying on manpath, which exists on freebsd and
with man-db, but not mandoc. Use `man -w` instead, which works
with both mandoc and man-db. Also do nothing if man does not
exist, just do not search manpaths then, e.g. in environments
where base-man is not installed.