main/initrmafs-tools: fix up symlinking in hooks
The relpath function would previously create entirely wrong
symlink targets, which just happened to work thanks to usrmerge
and some accidental misbehavior.

While at it, also just use copy_exec to get in our core tools,
which will take care of stuff like libxo and so on so we do not
have to copy them explicitly.